stats-yahoo/requirements.txt
yahoofantasy==1.4.8
pandas>=1.5.0
python-dotenv>=0.19.0
requests>=2.28.0
